+++ lib/utils/ym_date_util.dart
@@ -33,9 +33,6 @@ class YmDateUtil {
     int year = dateTime.year;
     int month = dateTime.month;
     int day = dateTime.day;
-    int hour = dateTime.hour;
-    int minute = dateTime.minute;
-    int second = dateTime.second;
     return format2Number(year.toString()) + "-" + format2Number(month.toString()) + "-" + format2Number(day.toString());
   }